+++ ada/tests/src/ams_mel_smoke.adb
@@ -12,6 +12,7 @@ with AMS_MEL_IR_C2_Metadata_Tests;
 with AMS_MEL_IR_C2_Common_Tests;
 with AMS_MEL_IR_Health_Status_Tests;
 with AMS_MEL_IR_Instrumentation_Tests;
+with AMS_MEL_IR_Track_Tests;
 with GNAT.OS_Lib;
 
 procedure AMS_MEL_Smoke is
@@ -222,6 +223,7 @@ begin
    AMS_MEL_IR_C2_Common_Tests.Run (Provider_Path);
    AMS_MEL_IR_Health_Status_Tests.Run (Provider_Path);
    AMS_MEL_IR_Instrumentation_Tests.Run (Provider_Path);
+   AMS_MEL_IR_Track_Tests.Run (Provider_Path);
    Cleanup_Lifetime_Log;
    Ada.Text_IO.Put_Line ("PASS: Ada provider load/init/version/close/finalization contract");
 exception